Comments: Rename package assertion -> upgradecheck
